Boring cleanup
zzz

+++ include/3dstris/states/sprinttimes.hpp
@@ -4,7 +4,6 @@
 #include <3dstris/gui/panel.hpp>
 #include <3dstris/state.hpp>
 #include <3dstris/util/text.hpp>
-#include <vector>
 
 class SprintTimes : public State {
    public:
@@ -16,19 +15,19 @@ class SprintTimes : public State {
 	void genValues();
 
    private:
-	static const u32 CELLS = 3;
-	static const u32 CELL_H = 40;
-	static const u32 TABLE_W = 300;
-	static const u32 TABLE_H = (CELLS + 1) * CELL_H;
-	static const u32 TABLE_X = (SCREEN_WIDTH - TABLE_W) / 2;
-	static const u32 TABLE_Y = (SCREEN_HEIGHT - TABLE_H) / 2;
+	static const u8 CELLS = 3;
+	static const u16 CELL_H = 40;
+	static const u16 TABLE_W = 300;
+	static const u16 TABLE_H = (CELLS + 1) * CELL_H;
+	static const u16 TABLE_X = (SCREEN_WIDTH - TABLE_W) / 2;
+	static const u16 TABLE_Y = (SCREEN_HEIGHT - TABLE_H) / 2;
 
 	GUI gui;
 	Panel panel;
 	Button& backButton;
 
-	static const u32 TIME_W = TABLE_W * 2 / 5;
-	static const u32 DATE_W = TABLE_W - TIME_W;
+	static const u16 TIME_W = TABLE_W * 2 / 5;
+	static const u16 DATE_W = TABLE_W - TIME_W;
 
 	Text timeLabel;
 	Text dateLabel;
